--- clients/cli/src/config/mod.rs
@@ -1,8 +1,0 @@
-//! Configuration management for the application.
-//!
-//! This module handles various configuration aspects:
-//! * `environment` - Environment-specific configuration (analytics, web socket URLs)
-//! * `prover` - Prover-specific configuration and identity management
-
-pub mod environment;
-pub mod prover;
