--- clients/cli/src/network/mod.rs
@@ -1,7 +1,0 @@
-//! Network-related functionality for handling communications.
-//!
-//! This module contains components for managing network connections
-//! and websocket communications.
-
-pub mod connection;
-pub mod websocket;
