--- clients/cli/src/network/websocket.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-mod tests {}
