+++ clients/cli/src/prover.rs
@@ -1,17 +1,18 @@
 // Copyright (c) 2024 Nexus. All rights reserved.
 
+mod analytics;
 mod config;
+mod connection;
 mod generated;
-mod network;
-mod utils;
+mod prover_id_manager;
+mod websocket;
 
-use crate::utils::analytics::track;
+use crate::analytics::track;
 
 use std::borrow::Cow;
 
-use crate::network::connection::connect_to_orchestrator_with_retry;
+use crate::connection::connect_to_orchestrator_with_retry;
 
-use crate::config::prover::ProverConfig;
 use clap::Parser;
 use futures::SinkExt;
 use generated::pb::ClientProgramProofRequest;
@@ -19,6 +20,7 @@ use prost::Message as _;
 use serde_json::json;
 use std::time::Instant;
 // Network connection types for WebSocket communication
+
 // WebSocket protocol types for message handling
 use tokio_tungstenite::tungstenite::protocol::{
     frame::coding::CloseCode, // Status codes for connection closure (e.g., 1000 for normal)
@@ -35,11 +37,7 @@ use nexus_core::{
         NexusVM,
     },
     prover::nova::{
-        init_circuit_trace,
-        key::CanonicalSerialize,
-        // pp::gen_vm_pp,
-        prove_seq_step,
-        types::IVCProof, // types::*,
+        init_circuit_trace, key::CanonicalSerialize, pp::gen_vm_pp, prove_seq_step, types::*,
     },
 };
 use std::fs;
@@ -72,10 +70,7 @@ fn get_file_as_byte_vec(filename: &str) -> Vec<u8> {
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    //
-    // 1. INITIALIZATION AND SETUP
-    // Configure tracing, parse arguments, and load prover configuration
-    //
+    // Configure the tracing subscriber
     tracing_subscriber::fmt()
         .with_env_filter(EnvFilter::from_default_env())
         .with_span_events(FmtSpan::CLOSE)
@@ -83,18 +78,20 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let args = Args::parse();
 
-    let ProverConfig {
-        ws_addr_string,
-        k,
-        prover_id,
-        public_parameters,
-    } = config::prover::initialize(args.hostname, args.port).await?;
+    let ws_addr_string = format!(
+        "{}://{}:{}/prove",
+        if args.port == 443 { "wss" } else { "ws" },
+        args.hostname,
+        args.port
+    );
 
-    //
-    // 2. NETWORK SETUP
-    // Establish WebSocket connection to orchestrator with retry logic
-    //
-    let mut client = connect_to_orchestrator_with_retry(&ws_addr_string, &prover_id).await;
+    let k = 4;
+    // TODO(collinjackson): Get parameters from a file or URL.
+    let pp = gen_vm_pp::<C1, seq::SetupParams<(G1, G2, C1, C2, RO, SC)>>(k as usize, &())
+        .expect("error generating public parameters");
+
+    // get or generate the prover id
+    let prover_id = prover_id_manager::get_or_generate_prover_id();
 
     track(
         "connect".into(),
@@ -103,29 +100,30 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         json!({"prover_id": prover_id}),
     );
 
-    //
-    // 3. MAIN PROVING LOOP
-    // Continuously receive programs, generate and send proofs
-    //
+    // Connect to the Orchestrator with exponential backoff
+    let mut client = connect_to_orchestrator_with_retry(&ws_addr_string, &prover_id).await;
+
+    track(
+        "register".into(),
+        format!("Your assigned prover identifier is {}.", prover_id),
+        &ws_addr_string,
+        json!({"ws_addr_string": ws_addr_string, "prover_id": prover_id}),
+    );
     loop {
-        // Generate random input data for the program
-        use rand::Rng;
+        // Create the inputs for the program
+        use rand::Rng; // Required for .gen() methods
         let mut rng = rand::thread_rng();
         let input = vec![5, rng.gen::<u8>(), rng.gen::<u8>()];
 
-        // Initialize the Nexus Virtual Machine with the program (fast-fib)
-        // and set its input parameters
         let mut vm: NexusVM<MerkleTrie> =
             parse_elf(get_file_as_byte_vec("src/generated/fast-fib").as_ref())
                 .expect("error loading and parsing RISC-V instruction");
         vm.syscalls.set_input(&input);
 
         // TODO(collinjackson): Get outputs
-        // Generate execution trace of the program
         let completed_trace = trace(&mut vm, k as usize, false).expect("error generating trace");
         let tr = init_circuit_trace(completed_trace).expect("error initializing circuit trace");
 
-        // Configure proving parameters
         let total_steps = tr.steps();
         let start = 0;
         let steps_to_prove = 10;
@@ -134,7 +132,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             end = total_steps
         }
 
-        // Initialize the proof with starting state
         let z_st = tr.input(start).expect("error starting circuit trace");
         let mut proof = IVCProof::new(&z_st);
 
@@ -160,26 +157,22 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         let start_time = Instant::now();
         let mut progress_time = start_time;
         for step in start..end {
-            // Generate proof for current step
-            proof =
-                prove_seq_step(Some(proof), &public_parameters, &tr).expect("error proving step");
+            proof = prove_seq_step(Some(proof), &pp, &tr).expect("error proving step");
             steps_proven += 1;
             completed_fraction = steps_proven as f32 / steps_to_prove as f32;
 
-            // Calculate performance metrics
             let progress_duration = progress_time.elapsed();
             let proof_cycles_hertz = k as f64 * 1000.0 / progress_duration.as_millis() as f64;
 
-            // Send progress update to orchestrator
             let progress = ClientProgramProofRequest {
                 steps_in_trace: total_steps as i32,
                 steps_proven,
                 step_to_start: start as i32,
+                program_id: "fast-fib".to_string(),
+                client_id_token: None,
                 proof_duration_millis: progress_duration.as_millis() as i32,
                 k,
                 cli_prover_id: Some(prover_id.clone()),
-                program_id: "fast-fib".to_string(),
-                client_id_token: None,
             };
 
             track(
@@ -223,7 +216,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 tokio::time::sleep(tokio::time::Duration::from_secs(u64::pow(2, retries))).await;
             }
 
-            // On final step, compress and store the proof
             if step == end - 1 {
                 let mut buf = Vec::new();
                 let mut writer = Box::new(&mut buf);
@@ -244,10 +236,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         }
     }
 
-    //
-    // 4. CLEANUP AND SHUTDOWN
-    // Clean termination of WebSocket connection
-    //
     client
         .close(Some(CloseFrame {
             code: CloseCode::Normal,
