--- clients/cli/src/utils/mod.rs
@@ -1,8 +1,0 @@
-//! Utility functions and helpers used throughout the application.
-//!
-//! This module contains various utility components:
-//! * `analytics` - Analytics and metrics tracking utilities
-//! * `prover_id` - Prover identity generation and management
-
-pub mod analytics;
-pub mod prover_id;
