+++ clients/cli/src/websocket.rs
@@ -1,0 +1,3 @@
+#[cfg(test)]
+mod tests {
+}
