+++ proto/orchestrator.proto
@@ -57,6 +57,26 @@ enum Network {
   NETWORK_MAINNET = 3;
 }
 
+message ProofRequest {
+  CompiledProgram program = 1;
+
+  VMProgramInput input = 2;
+
+  // Step of the trace to start the proof, inclusive.
+  //
+  // If missing, proving starts at the beginning of the trace.
+  optional int32 step_to_start = 3;
+
+  // Number of steps for this proof request.
+  //
+  // If zero, proving is skipped. If missing, all steps are proved.
+  optional int32 steps_to_prove = 4;
+}
+
+message ProofResponse {
+  Proof proof = 1;
+}
+
 // Send a request to the orchestrator backend to request the current leaderboard.
 message LeaderboardRequest {
   int32 offset = 1; // The offset of the leaderboard to request.
@@ -71,6 +91,7 @@ message LeaderboardResponse {
 // An entry in the leaderboard.
 message LeaderboardResponseEntry {
   string client_id = 1;
+  int64 prover_id = 2;
   int64 total_cycles = 3;
   double cycles_per_sec = 4;
   int64 rank = 5;
@@ -138,4 +159,97 @@ message ClientDataResponse {
     // Response for the total number of cycles proved by a user.
     UserCyclesProvedResponse user_cycles_proved = 4;
   }
+}
+
+// A message that always represents a program runnable on the Nexus VM.
+message CompiledProgram {
+  oneof program {
+    // ELF binary containing a program to be proved, expressed in the RV32I ISA.
+    bytes rv32i_elf_bytes = 1;
+  }
+}
+
+message VMProgramInput {
+  oneof input {
+    // Input expressed as raw bytes to be read as-is off of the input tape.
+    bytes raw_bytes = 1;
+  }
+}
+
+message Proof {
+  oneof proof {
+    bytes nova_bytes = 1;
+  }
+}
+
+message Progress {
+  // Completion status expressed as a number between zero and one,
+  // inclusive.
+  float completed_fraction = 1;
+
+  // The total size of the execution trace in steps.
+  int32 steps_in_trace = 2;
+
+  // The number of steps of the execution trace to be proven.
+  int32 steps_to_prove = 3;
+
+  // The number of steps proven so far.
+  int32 steps_proven = 4;
+}
+
+// Streamed messages sent to the orchestrator to keep it updated with the
+// prover's status.
+message ProverRequest {
+  oneof contents {
+    // Details about this supply node for use by the orchestrator.
+    ProverRequestRegistration registration = 1;
+
+    // A completed proof.
+    Proof proof = 2;
+
+    // Periodic progress update for the current proof.
+    Progress progress = 3;
+
+    // Periodic liveness indicator when no proof is being computed.
+    Heartbeat heartbeat = 4;
+  }
+}
+
+enum ProverType {
+  // Experimental new prover types should leave the prover type unspecified.
+  PROVER_TYPE_UNSPECIFIED = 0;
+
+  // The default prover type, used for volunteered compute resources.
+  PROVER_TYPE_VOLUNTEER = 1;
+
+  // Provers running on public continuous integration.
+  // May restrict the types of programs that can be assigned.
+  PROVER_TYPE_CI = 2;
+}
+
+// Metadata that helps the orchestrator schedule work to the requesting compute
+// supplier.
+message ProverRequestRegistration {
+  // What type of prover this is.
+  ProverType prover_type = 1;
+
+  // A unique identifier for this prover, generated by the prover.
+  //
+  // Distinct provers must not share an identifier; do not use a constant value.
+  string prover_id = 2;
+
+  // The number of proof cycles that this prover expects to compute
+  // over the course of one second. Proof cycles are proof steps times k.
+  optional double estimated_proof_cycles_hertz = 3;
+
+  // The network to prove on.
+  Network network = 4;
+}
+
+message ProverResponse {
+  // Forward the literal request for now
+  ProofRequest to_prove = 1;
+}
+
+message Heartbeat {
 }
